Refactor: split layer list into files
