+++ src/packages/toc/ui/LayerList/hooks.ts
@@ -1,0 +1,34 @@
+// SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
+// SPDX-License-Identifier: Apache-2.0
+import { AnyLayer, isSublayer, Layer, MapModel } from "@open-pioneer/map";
+import { useReactiveSnapshot } from "@open-pioneer/reactivity";
+
+/** Returns the top level operational layers in render order (topmost layer first). */
+export function useLayers(map: MapModel): Layer[] {
+    return useReactiveSnapshot(() => {
+        const layers = map.layers.getOperationalLayers({ sortByDisplayOrder: true }) ?? [];
+        layers.reverse(); // render topmost layer first
+        return layers;
+    }, [map]);
+}
+
+/**
+ * Returns the child layers (sublayers or layers contained in a group layer) of a layer.
+ * Layers are returned in render order (topmost sublayer first).
+ */
+export function useChildLayers(layer: AnyLayer): AnyLayer[] | undefined {
+    return useReactiveSnapshot(() => {
+        const children = layer.children?.getItems({ sortByDisplayOrder: true });
+        children?.reverse(); // render topmost layer first
+        return children;
+    }, [layer]);
+}
+
+/** Returns the layers current state. */
+export function useLoadState(layer: AnyLayer): string {
+    return useReactiveSnapshot(() => {
+        // for sublayers, use the state of the parent
+        const target = isSublayer(layer) ? layer.parentLayer : layer;
+        return target.loadState;
+    }, [layer]);
+}
