+++ src/packages/toc/utils/slug.ts
@@ -1,0 +1,14 @@
+// SPDX-FileCopyrightText: 2023 Open Pioneer project (https://github.com/open-pioneer)
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * Strips special characters from the given string, making it suitable
+ * for use in CSS classes.
+ */
+export function slug(id: string): string {
+    return id
+        .toLowerCase()
+        .replace(/[^a-z0-9 -]/g, "")
+        .replace(/\s+/g, "-")
+        .replace(/-+/g, "-");
+}
